backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.1.18
celery==5.3.6
redis==5.0.3
gunicorn==21.2.0
django-cors-headers==4.3.1
django-environ==0.11.2
whitenoise==6.6.0
django-redis==5.4.0
